Support adapting core wasm to components (#374)
This PR adds a component adapter based on the wasi_snapshot_preview1 adapter
that also supports the fastly host call abi. The adapter can be used in two
forms:

1. using the new viceroy adapt command to adapt a core-wasm module to a
   component ahead-of-time,
2. using the --adapt flag to viceroy serve or viceroy run, adapting the
   core-wasm in-memory. This functionality opens the door to running viceroy's
   existing test suite with components, and also to a longer-term migration
   towards components in-general.

This feature comes with a downside to the development experience though: any
changes made to the adapter require that we also run the make adapter target,
and commit the changes to the wasi_snapshot_preview1.wasm binary. It would be
great to check if this was necessary in CI, but it's not clear to me what the
right mechanism is to enforce that. However, changes to the adapter should be
infrequent, so hopefully this isn't too much of a burden.