🎁 Update to wasmtime 22.0.0 (#392)
Update to wasmtime 22.0.0.

This included some pretty substantial changes to the wiggle crate, which removes
runtime borrow checking of guest memory. This is a huge ergonomic improvement,
as the borrows are all tracked as rrust borrows instead.

One piece of functionality that's missing after the wiggle changes is the
ability to interact with a slice of guest memory that's not &[u8]. To work
around this, I'm copying memory over to the host, but in future releases of
wiggle we should have the ability to interact with slices that aren't just to
u8.